+++ lldb/source/Symbol/CMakeLists.txt
@@ -14,7 +14,6 @@ add_lldb_library(lldbSymbol
   CompilerType.cpp
   DWARFCallFrameInfo.cpp
   DebugMacros.cpp
-  Declaration.cpp
   DeclVendor.cpp
   FuncUnwinders.cpp
   Function.cpp
